requirements.txt
pandas
numpy
psycopg2-binary
websockets
aiohttp
pyarrow
scikit-learn
joblib
torch
tqdm 